requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3
gunicorn==21.2.0
pyopenssl>=25.0.0
waitress>=3.0.0
onnxruntime>=1.16.0
pillow>=10.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
